+++ src/proxy/planckProxy.py
@@ -8,9 +8,10 @@ import logging
 import os
 
 # Create loggers
-console_logger = logging.getLogger('consoleLogger')
-file_logger = logging.getLogger('fileLogger')
+# console_logger = logging.getLogger('consoleLogger')
+# file_logger = logging.getLogger('fileLogger')
 
+from proxy.utils.sysloggers import console_logger, file_logger
 from proxy.utils.message import Message
 from proxy.utils.hooks import cleanup, except_hook
 from proxy.utils.printers import dbg, print_init_info, print_summary_info, print_keys_and_headers
